+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 Note: Numbers like (#123) point to closed Pull Requests on the fractal-vizarr-viewer repository.
 
+# Unreleased
+
+* Update `User` type (\#74);
+* Drop `TestingBasicAuthAuthorizer` and `NoneAuthorizer` (\#75);
+
 # 0.4.2
 
 * Added workaround to fix labels bug in most recent Chrome versions (\#68);
-* Update `User` type (\#74);
 
 # 0.4.0