+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Note: Numbers like (#123) point to closed Pull Requests on the fractal-vizarr-viewer repository.
 
-# Unreleased
+# 0.4.3
 
 * Update `User` type (\#74);
 * Drop `TestingBasicAuthAuthorizer` and `NoneAuthorizer` (\#75);